Treat treat string log messages as dicts if they are in json format (#119)
Due to the nature of how RodatingFileHandler is implemented, the log recors are formatted twice.
Once in `def shouldRollover(self, record):` in `class RotatingFileHandler` and second time when `emit` is actually called after that
https://github.com/python/cpython/blob/master/Lib/logging/handlers.py#L71

That results in `ProcessorFormatter` formatting the same message twice, but the second time, it is not aware if the message itslef is a dict, as it arrives as string.

Here we are handling exactly this case - we check if the string record message that arrives, has json structure, and if it does, we treat it, as if we would treat a dict