Add docs around using ProcessorFormatter (#140)
This adds a (I think) gentler introduction to the way that ProcessorFormatter
works. The existing dictConfig remains. I wasn't sure if I should maybe add
another couple of headings to the docs to make it more clear, we're already at
the maximum nesting level in sphinx so it won't show up in the ToC anyway.

I think that if this had existed I wouldn't have opened #138