Correctly figure out exc_info in ConsoleRenderer.
If the `exc_info` key was filled with a raw exception, and not a tuple,
the `ConsoleRenderer` code would call `sys.exc_info` and overwrite it.
This is different to how the other processors handle the exc_info key.

Additionally, if that call to `sys.exc_info` returned None, an explosion
would happen.  This is the case if we placed an exception in `exc_info`,
but did not actually log until outside of the `except` block.