Allow CallsiteParameterAdder to be pickled
Prior to this commit, the `CallsiteParameterAdder` object could not be
pickled.  As a result, structlog configurations could not be propagated
to subprocesses using `pickle.dumps(structlog.get_config())`.  This
commit updates the handlers in `CallsiteParameterAdder._handlers` to be
free functions, rather than lambda functions, so that all class members
can be pickled.

Closes #600