Allow CallsiteParameterAdder to be pickled (#603)
* Allow CallsiteParameterAdder to be pickled

Prior to this commit, the `CallsiteParameterAdder` object could not be
pickled.  As a result, structlog configurations could not be propagated
to subprocesses using `pickle.dumps(structlog.get_config())`.  This
commit updates the handlers in `CallsiteParameterAdder._handlers` to be
free functions, rather than lambda functions, so that all class members
can be pickled.

Closes #600

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Updated changelog with link to PR

* Update with type annotations, correct function naming conventions

* Update tests/processors/test_processors.py

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Hynek Schlawack <hs@ox.cx>