Fix reference to undefined var `logger` in "Getting Started" example
While following along with this "Getting Started" guide for the first
time, I got sidetracked for several minutes by this `logger`.

None of the example snippets earlier in this file assigned to a variable
named `logger` (but they frequently assigned the result of
`structlog.get_logger()` to a variable named `log`).

I spent a bit poking around in the API reference to try and figure out
which types define a `.bind()` method and whether they completely
overlap with the types returned by `structlog.get_logger()`.  

Eventually I decided the specific variable name probably wasn't a super
considered choice and I was overthinking things.

Maybe this change could save someone time in the future.